streamlit_app/requirements.txt
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
trimesh==4.0.5
